Use readqueue for results and change ownership
