+++ photon-lib/src/main/native/cpp/photon/PhotonPoseEstimator.cpp
@@ -66,22 +66,6 @@ PhotonPoseEstimator::PhotonPoseEstimator(frc::AprilTagFieldLayout tags,
                                          frc::Transform3d robotToCamera)
     : aprilTags(tags),
       strategy(strat),
-      camera(nullptr),
-      m_robotToCamera(robotToCamera),
-      lastPose(frc::Pose3d()),
-      referencePose(frc::Pose3d()),
-      poseCacheTimestamp(-1_s) {
-  HAL_Report(HALUsageReporting::kResourceType_PhotonPoseEstimator,
-             InstanceCount);
-  InstanceCount++;
-}
-
-PhotonPoseEstimator::PhotonPoseEstimator(frc::AprilTagFieldLayout tags,
-                                         PoseStrategy strat, PhotonCamera&& cam,
-                                         frc::Transform3d robotToCamera)
-    : aprilTags(tags),
-      strategy(strat),
-      camera(std::make_shared<PhotonCamera>(std::move(cam))),
       m_robotToCamera(robotToCamera),
       lastPose(frc::Pose3d()),
       referencePose(frc::Pose3d()),
@@ -104,25 +88,6 @@ void PhotonPoseEstimator::SetMultiTagFallbackStrategy(PoseStrategy strategy) {
     InvalidatePoseCache();
   }
   multiTagFallbackStrategy = strategy;
-}
-
-std::optional<EstimatedRobotPose> PhotonPoseEstimator::Update() {
-  if (!camera) {
-    FRC_ReportError(frc::warn::Warning, "[PhotonPoseEstimator] Missing camera!",
-                    "");
-    return std::nullopt;
-  }
-  auto result = camera->GetLatestResult();
-  return Update(result, camera->GetCameraMatrix(), camera->GetDistCoeffs());
-}
-
-std::optional<EstimatedRobotPose> PhotonPoseEstimator::Update(
-    const PhotonPipelineResult& result) {
-  // If camera is null, best we can do is pass null calibration data
-  if (!camera) {
-    return Update(result, std::nullopt, std::nullopt, this->strategy);
-  }
-  return Update(result, camera->GetCameraMatrix(), camera->GetDistCoeffs());
 }
 
 std::optional<EstimatedRobotPose> PhotonPoseEstimator::Update(
@@ -399,6 +364,10 @@ std::optional<EstimatedRobotPose> PhotonPoseEstimator::MultiTagOnRioStrategy(
   }
 
   if (!camMat || !distCoeffs) {
+    FRC_ReportError(frc::warn::Warning,
+                    "No camera calibration data provided to "
+                    "PhotonPoseEstimator::MultiTagOnRioStrategy!",
+                    "");
     return Update(result, std::nullopt, std::nullopt,
                   this->multiTagFallbackStrategy);
   }
