+++ photon-lib/src/test/java/org/photonvision/PhotonPoseEstimatorTest.java
@@ -125,10 +125,9 @@ class PhotonPoseEstimatorTest {
         cameraOne.result.setTimestampSeconds(11);
 
         PhotonPoseEstimator estimator =
-                new PhotonPoseEstimator(
-                        aprilTags, PoseStrategy.LOWEST_AMBIGUITY, cameraOne, new Transform3d());
+                new PhotonPoseEstimator(aprilTags, PoseStrategy.LOWEST_AMBIGUITY, new Transform3d());
 
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         Pose3d pose = estimatedPose.get().estimatedPose;
 
         assertEquals(11, estimatedPose.get().timestampSeconds);
@@ -208,10 +207,9 @@ class PhotonPoseEstimatorTest {
                 new PhotonPoseEstimator(
                         aprilTags,
                         PoseStrategy.CLOSEST_TO_CAMERA_HEIGHT,
-                        cameraOne,
                         new Transform3d(new Translation3d(0, 0, 4), new Rotation3d()));
 
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         Pose3d pose = estimatedPose.get().estimatedPose;
 
         assertEquals(4, estimatedPose.get().timestampSeconds);
@@ -290,11 +288,10 @@ class PhotonPoseEstimatorTest {
                 new PhotonPoseEstimator(
                         aprilTags,
                         PoseStrategy.CLOSEST_TO_REFERENCE_POSE,
-                        cameraOne,
                         new Transform3d(new Translation3d(0, 0, 0), new Rotation3d()));
         estimator.setReferencePose(new Pose3d(1, 1, 1, new Rotation3d()));
 
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         Pose3d pose = estimatedPose.get().estimatedPose;
 
         assertEquals(17, estimatedPose.get().timestampSeconds);
@@ -373,12 +370,11 @@ class PhotonPoseEstimatorTest {
                 new PhotonPoseEstimator(
                         aprilTags,
                         PoseStrategy.CLOSEST_TO_LAST_POSE,
-                        cameraOne,
                         new Transform3d(new Translation3d(0, 0, 0), new Rotation3d()));
 
         estimator.setLastPose(new Pose3d(1, 1, 1, new Rotation3d()));
 
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         Pose3d pose = estimatedPose.get().estimatedPose;
 
         cameraOne.result =
@@ -444,7 +440,7 @@ class PhotonPoseEstimatorTest {
                                                 new TargetCorner(7, 8)))));
         cameraOne.result.setTimestampSeconds(7);
 
-        estimatedPose = estimator.update();
+        estimatedPose = estimator.update(cameraOne.result);
         pose = estimatedPose.get().estimatedPose;
 
         assertEquals(7, estimatedPose.get().timestampSeconds);
@@ -485,25 +481,24 @@ class PhotonPoseEstimatorTest {
                 new PhotonPoseEstimator(
                         aprilTags,
                         PoseStrategy.AVERAGE_BEST_TARGETS,
-                        cameraOne,
                         new Transform3d(new Translation3d(0, 0, 0), new Rotation3d()));
 
         // Empty result, expect empty result
         cameraOne.result = new PhotonPipelineResult();
         cameraOne.result.setTimestampSeconds(1);
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         assertFalse(estimatedPose.isPresent());
 
         // Set actual result
         cameraOne.result = result;
-        estimatedPose = estimator.update();
+        estimatedPose = estimator.update(cameraOne.result);
         assertTrue(estimatedPose.isPresent());
         assertEquals(20, estimatedPose.get().timestampSeconds, .01);
         assertEquals(20, estimator.poseCacheTimestampSeconds);
 
         // And again -- pose cache should mean this is empty
         cameraOne.result = result;
-        estimatedPose = estimator.update();
+        estimatedPose = estimator.update(cameraOne.result);
         assertFalse(estimatedPose.isPresent());
         // Expect the old timestamp to still be here
         assertEquals(20, estimator.poseCacheTimestampSeconds);
@@ -513,7 +508,7 @@ class PhotonPoseEstimatorTest {
         assertEquals(-1, estimator.poseCacheTimestampSeconds);
         // Update should cache the current timestamp (20) again
         cameraOne.result = result;
-        estimatedPose = estimator.update();
+        estimatedPose = estimator.update(cameraOne.result);
         assertEquals(20, estimatedPose.get().timestampSeconds, .01);
         assertEquals(20, estimator.poseCacheTimestampSeconds);
     }
@@ -588,10 +583,9 @@ class PhotonPoseEstimatorTest {
                 new PhotonPoseEstimator(
                         aprilTags,
                         PoseStrategy.AVERAGE_BEST_TARGETS,
-                        cameraOne,
                         new Transform3d(new Translation3d(0, 0, 0), new Rotation3d()));
 
-        Optional<EstimatedRobotPose> estimatedPose = estimator.update();
+        Optional<EstimatedRobotPose> estimatedPose = estimator.update(cameraOne.result);
         Pose3d pose = estimatedPose.get().estimatedPose;
 
         assertEquals(20, estimatedPose.get().timestampSeconds, .01);
@@ -606,6 +600,11 @@ class PhotonPoseEstimatorTest {
         }
 
         PhotonPipelineResult result;
+
+        @Override
+        public List<PhotonPipelineResult> getAllUnreadResults() {
+            return List.of(result);
+        }
 
         @Override
         public PhotonPipelineResult getLatestResult() {
